Test: Rearrange tests for lgr_tasks
