+++ src/lgr_tasks/urls.py
@@ -1,10 +1,9 @@
-# -*- coding: utf-8 -*-
 from django.urls import path
 
-from .views import ProcessListView, DeleteProcessView, DeleteAllFinishedProcessView
+from lgr_tasks.views import DeleteAllFinishedProcessesView, DeleteProcessView, ProcessListView
 
 urlpatterns = [
     path('list', ProcessListView.as_view(), name='list_process'),
     path('<int:task_id>/delete', DeleteProcessView.as_view(), name='delete_process'),
-    path('delete', DeleteAllFinishedProcessView.as_view(), name='delete_finished'),
+    path('delete', DeleteAllFinishedProcessesView.as_view(), name='delete_finished'),
 ]
